requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsons==1.6.3
jwt==1.3.1
MarkupSafe==2.1.5
phonepe==1.0.1
pillow==10.3.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pymongo==4.6.3
python-dateutil==2.9.0.post0
razorpay==1.4.2
requests==2.31.0
s3transfer==0.10.1
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
typing_extensions==4.11.0
typish==1.9.3
urllib3==2.2.1
Werkzeug==3.0.2
